fix deadlock-prone API
the previous APIs expect &Task, which imply the caller should obtain the
task references from the concurrent hashmap before calling the APIs.

When a Task is first obtained through dashmap, and the called API
attemps to obtain this task through dashmap again, a deadlock occurs.

We make deadlocks much less likely by only requiring task IDs in APIs.
To make task IDs easier to deal with, we make them copyable by changing
them from strings to be u64s.

Ensuring TaskIds are unique is a separate issue that needs to be dealt
with.